requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.3
billiard==3.6.4.0
boto3==1.17.84
botocore==1.20.112
cached-property==1.5.2
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
codeless-django==0.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
defusedxml==0.7.1
dj-rest-auth==3.0.0
Django==3.2.18
django-adminlte-3==0.1.6
django-allauth==0.52.0
django-celery-beat==2.2.0
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-filter==22.1
django-formtools==2.4
django-otp==1.1.6
django-phonenumber-field==6.4.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-simple-history==3.3.0
django-storages==1.11.1
django-timezone-field==4.2.3
django-two-factor-auth==1.15.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
idna==3.4
importlib-metadata==6.6.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==0.10.0
kombu==5.2.4
MarkupSafe==2.1.2
oauthlib==3.2.2
packaging==23.0
Pillow==9.4.0
prompt-toolkit==3.0.38
psycopg2==2.9.7
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==1.7.1
pypng==0.20220715.0
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2022.7.1
qrcode==7.4.2
redis==5.0.0
requests==2.28.2
requests-oauthlib==1.3.1
rest-framework-simplejwt==0.0.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
zipp==3.15.0
